customer profile shows up and add a payment works
